+++ src/sql/src/plan/query.rs
@@ -2842,7 +2842,9 @@ fn plan_expr_inner<'a>(
         Expr::Row { exprs } => plan_row(ecx, exprs),
 
         // Generalized functions, operators, and casts.
-        Expr::Op { op, expr1, expr2 } => Ok(plan_op(ecx, op, expr1, expr2.as_deref())?.into()),
+        Expr::Op { op, expr1, expr2 } => {
+            Ok(plan_op(ecx, op.op_str(), expr1, expr2.as_deref())?.into())
+        }
         Expr::Cast { expr, data_type } => plan_cast(ecx, expr, data_type),
         Expr::Function(func) => Ok(plan_function(ecx, func)?.into()),